PQ 0.0.61: Several fixes: - Added battle log - Fixed max stage corrupt command after winning it.

+++ IdleHeroes.Bot/Services/IStageService.cs
@@ -5,6 +5,6 @@ namespace IdleHeroes.Services
 {
     public interface IStageService
     {
-        Task<Stage> GetStageFromNumber(ulong stageNumber);
+        Task<Stage> GetStageFromNumber(double stageNumber);
     }
 }
