+++ IdleHeroes.DAL/Models/Stage.cs
@@ -6,15 +6,15 @@ namespace IdleHeroesDAL.Models
 {
     public class Stage : Entity
     {
-        public ulong Number { get; set; }
-        public ulong XPPerMinute { get; set; }
-        public ulong CoinsPerMinute { get; set; }
-        public ulong FoodChancePerMinute { get; set; }
-        public ulong FoodAmount { get; set; }
-        public ulong GemsDropChancePerMinute { get; set; }
-        public ulong GemsAmount { get; set; }
-        public ulong RelicsDropChancePerMinute { get; set; }
-        public ulong RelicsAmount { get; set; }
+        public double Number { get; set; }
+        public double XPPerMinute { get; set; }
+        public double CoinsPerMinute { get; set; }
+        public double FoodChancePerMinute { get; set; }
+        public double FoodAmount { get; set; }
+        public double GemsDropChancePerMinute { get; set; }
+        public double GemsAmount { get; set; }
+        public double RelicsDropChancePerMinute { get; set; }
+        public double RelicsAmount { get; set; }
         public TimeSpan TimeToBeat { get; set; }
         public StageDifficultyEnum Difficulty { get; set; }
         public List<StageEnemy> Enemies { get; set; }
