+++ IdleHeroes.DAL/Models/TavernCompanion.cs
@@ -5,6 +5,6 @@ namespace IdleHeroesDAL.Models
     public class TavernCompanion : Entity
     {
         public Companion Companion { get; set; }
-        public ulong FoodCost { get; set; }
+        public double FoodCost { get; set; }
     }
 }
